Refactor project structure into one more HACS developer friendly
