+++ .github/FUNDING.yml
@@ -1,0 +1,2 @@
+github: [AnthonMS]
+custom: ['anthon's paypal', 'anthon's bymeacoffee']
