+++ .github/ISSUE_TEMPLATE/bug_report.yml
@@ -1,0 +1,52 @@
+name: 🐞 Bug Report
+description: Create a report to help us improve
+title: "[Bug]: <title>"
+labels: ["bug", "triage"]
+body:
+- type: textarea
+  attributes:
+    label: Current Behavior
+    description: A concise description of what you're experiencing.
+  validations:
+    required: false
+- type: textarea
+  attributes:
+    label: Expected Behavior
+    description: A concise description of what you expected to happen.
+  validations:
+    required: false
+- type: textarea
+  attributes:
+    label: Steps To Reproduce
+    description: Steps to reproduce the behavior.
+    placeholder: |
+      1. In this environment...
+      2. With this config...
+      3. Run '...'
+      4. See error...
+  validations:
+    required: false
+- type: textarea
+  attributes:
+    label: Environment
+    description: |
+      examples:
+        - **OS**: Ubuntu 20.04
+        - **Brower**: chrome
+        - **Version**: 1.6.2
+        - **Hassio**: 2022.2.3
+    value: |
+        - OS:
+        - Node:
+        - Version:
+        - Hassio:
+    render: markdown
+  validations:
+    required: false
+- type: textarea
+  attributes:
+    label: Anything else?
+    description: |
+      Links? References? Anything that will give us more context about the issue you are encountering!
+
+      Tip: You can attach images or log files by clicking this area to highlight it and then dragging files in.
