+++ .github/ISSUE_TEMPLATE/feature_request.yml
@@ -1,0 +1,33 @@
+name: 🚀 Feature request
+description: Suggest an idea for this project
+title: "[Feature]: <title>"
+labels: ["enhancement"]
+body:
+- type: textarea
+  attributes:
+    label: Is your feature request related to a problem?
+    description: >
+      A clear and concise description of what the problem is. Ex. I'm always frustrated when [...]
+  validations:
+    required: true
+- type: textarea
+  attributes:
+    label: Describe the solution you'd like
+    description: >
+      A clear and concise description of what you want to happen.
+  validations:
+    required: true
+- type: textarea
+  attributes:
+    label: Describe alternatives you've considered
+    description: >
+      A clear and concise description of any alternative solutions or features you've considered.
+- type: textarea
+  attributes:
+    label: Additional context
+    description: >
+      Add any other context or screenshots about the feature request.
+- type: markdown
+  attributes:
+    value: >
+      Thanks for contributing 🎉
