+++ .github/workflows/build.yml
@@ -1,0 +1,14 @@
+name: Build
+on:
+    pull_request:
+    push:
+jobs:
+    build:
+        runs-on: ubuntu-latest
+        steps:
+            - name: Checkout
+              uses: actions/checkout@v2
+            - name: Install
+              run: npm ci
+            - name: Build
+              run: npm run build
