+++ .github/workflows/validate.yaml
@@ -1,0 +1,15 @@
+name: HACS validation
+on:
+    push:
+        branches:
+            - main
+jobs:
+    validate:
+        runs-on: ubuntu-latest
+        steps:
+            - name: Checkout
+              uses: actions/checkout@v2
+            - name: HACS validation
+              uses: "hacs/action@main"
+              with:
+                  category: "plugin"
