+++ docs/cards/slider.md
@@ -20,10 +20,10 @@ Everything under function in the options table is css styling, so it is possible
 
 ### My Setup Using This
 
-![Example Setup](/other/docs/captures/my-slider/my-use-case.JPG)
+![Example Setup](/docs/images/my-slider/my-use-case.JPG)
 
 ### Demo
-![Example](/other/docs/captures/my-slider/ha-slider-card-demo.gif)
+![Example](/docs/images/my-slider/ha-slider-card-demo.gif)
 
 ## Options
 | Name | Type | Default | Description |
@@ -94,7 +94,7 @@ cards:
     step: "10"
 ```
 
-![Default Slider Config](/other/docs/captures/my-slider/default.JPG)
+![Default Slider Config](/docs/images/my-slider/default.JPG)
 
 #### Change Slider Height and Colors
 
@@ -107,7 +107,7 @@ cards:
       secondarySliderColor: 'red'
 ```
 
-![Change Slider Height and Colors Config](/other/docs/captures/my-slider/colors-height.JPG)
+![Change Slider Height and Colors Config](/docs/images/my-slider/colors-height.JPG)
 
 #### Change Thumb Size and Colors
 
@@ -123,7 +123,7 @@ cards:
     thumbColor: 'black'
 ```
 
-![Change Thumb Height and Colors Config](/other/docs/captures/my-slider/thumb-colors-height.JPG)
+![Change Thumb Height and Colors Config](/docs/images/my-slider/thumb-colors-height.JPG)
 
 #### Change Thumb Padding and Size
 
@@ -155,7 +155,7 @@ cards:
 
 ```
 
-![Change Thumb Height and Colors Config](/other/docs/captures/my-slider/thumb-padding-2.JPG)
+![Change Thumb Height and Colors Config](/docs/images/my-slider/thumb-padding-2.JPG)
 
 #### Change Active and Inactive Colors Config
 
@@ -171,7 +171,7 @@ cards:
     thumbColorOff: black
 ```
 
-![Change Active and Inactive Colors Config](/other/docs/captures/my-slider/thumb-colors-change.gif)
+![Change Active and Inactive Colors Config](/docs/images/my-slider/thumb-colors-change.gif)
 
 #### Using a `Switch` or `Lock`
 
@@ -187,7 +187,7 @@ cards:
 
 
 
-![Change Active and Inactive Colors Config](/other/docs/captures/my-slider/switch-lock-example.gif)
+![Change Active and Inactive Colors Config](/docs/images/my-slider/switch-lock-example.gif)
 ### NOTE
 
 When changing the padding of the thumb. If you want for example padding on the sides (horizontal) you will have to triple the width of the thumb itself, this has something to do with the border styling. There is possibly a way around this, if I use some more time on the styling, but for now, this will have to do, since it serves the purpose I need it for. Plus if you just keep this in mind, there should be no trouble. But play around with it. This includes when wanting padding on top/bottom (vertical), you will have to change height of the thumb.
