+++ docs/images/my-slider-v2/examples.png
