+++ docs/images/my-slider/ha-slider-card-demo.gif
