+++ docs/images/my-slider/my-use-case.JPG
