+++ docs/images/my-slider/switch-lock-example.gif
