+++ docs/images/my-slider/thumb-colors-change.gif
