+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "my-cards",
-  "version": "1.0.1",
+  "version": "2.0.0",
   "description": "Bundle of my custom Lovelace cards",
   "keywords": [
     "home-assistant",
